Issue #11163: Enforced file size inputs
